build: Allow LDFLAGS_FOR_TARGET to be passed to GCC configure script.
This patch is a small hack that allows the value of LDFLAGS_FOR_TARGET
as defined when the build configure script is invoked to be passed
also to the GCC build script.

This is necessary to build the x86 NDK toolchain. Probably not the
cleanest way to do it though, so the compiler team may want something
else for the sources on android.git.kernel.org.

Change-Id: I2b34ad6c28ede29a8b8244dd54ec607369661b5a